+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+#### Changed
+- The webapp viewer now uses PySide6 by default and PySide2 as a fallback (set QT_API=PySide2 to always use PySide2).
 
 ## [0.2.0] - 2022-04-01
 ### Added